replace a shared estimator
